refactor(qwen35): split scheduler event dispatch into scheduler/emit.rs
Extract terminal_scheduler_shutdown/prune_closed_requests/reject_unsupported_echo/
send_rejection (+ their echo-message const) out of scheduler.rs into a flat
scheduler/emit.rs submodule. Request-lifecycle TYPES stay in scheduler.rs because
steps.rs constructs them; emit.rs reaches them via use super::*. pub(super) +
use super::* as established. cargo check + fmt + test --lib (95/95) green.
